Check at the end of the e2e test cgroup cpu.max files have been reset

+++ test/e2e-kind.sh
@@ -32,6 +32,25 @@ else
     exit 1
 fi
 
+# also check that cgroup cpu.max file is back to the default limits
+python_with_boost_pod_uid=$(kubectl get pods python-with-boost -o jsonpath='{.metadata.uid}' | sed 's~-~_~g')
+python_with_boost_python_container_id=$(kubectl get pods python-with-boost -o jsonpath='{.status.containerStatuses[?(@.name=="python")].containerID}' | cut -d '/' -f 3)
+
+docker exec -it kind-worker cat /sys/fs/cgroup/kubelet.slice/kubelet-kubepods.slice/kubelet-kubepods-pod${python_with_boost_pod_uid}.slice/cpu.max > pod_cpu.max
+docker exec -it kind-worker cat /sys/fs/cgroup/kubelet.slice/kubelet-kubepods.slice/kubelet-kubepods-pod${python_with_boost_pod_uid}.slice/cri-containerd-${python_with_boost_python_container_id}.scope/cpu.max > python_container_cpu.max
+
+if ! diff -b <(cat pod_cpu.max) <(echo "5000 100000")
+then
+    echo -e "\033[0;31m[FAILURE]\033[0m pod cgroup cpu.max has not been reset"
+    exit 1
+fi
+
+if ! diff -b <(cat python_container_cpu.max) <(echo "5000 100000")
+then
+    echo -e "\033[0;31m[FAILURE]\033[0m python container cgroup cpu.max has not been reset"
+    exit 1
+fi
+
 kubectl delete \
     -f ./test/e2e/python-no-boost.yaml \
     -f ./test/e2e/python-with-boost.yaml
